final first version

+++ posts/views.py
@@ -4,9 +4,9 @@ from django.core.mail import send_mail
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 
+from billboard.settings import LOGIN_URL
 from .models import Post, Category, Reply
 from .forms import PostForm, ReplyForm
-from users.models import User
 
 
 class PostListView(ListView):
@@ -37,6 +37,12 @@ class PostCreateView(LoginRequiredMixin, CreateView):
         else:
             return self.form_invalid(form)
 
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['post_title'] = 'Создание объявления'
+        context['btn_name'] = 'Добавить'
+        return context
+
 
 class PostUpdateView(LoginRequiredMixin, UpdateView):
     template_name = 'posts/post_create.html'
@@ -52,23 +58,24 @@ class PostUpdateView(LoginRequiredMixin, UpdateView):
         id = self.kwargs.get('pk')
         return Post.objects.get(pk=id)
 
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['post_title'] = 'Редактировать объявление'
+        context['btn_name'] = 'Сохранить'
+        return context
+
 
 class PostDeleteView(LoginRequiredMixin, DeleteView):
     template_name = 'posts/post_delete.html'
     queryset = Post.objects.all()
     success_url = '/'
 
-    def delete(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        success_url = self.get_success_url()
-        print(self.object.author)
-        print(request.user)
-        if self.object.author == request.user:
-            self.object.delete()
-        return HttpResponseRedirect(success_url)
-
     def form_valid(self, form):
-        pass
+        self.object = self.get_object()
+        if self.object.author == self.request.user:
+            self.object.delete()
+            return HttpResponseRedirect(self.get_success_url())
+        return HttpResponseRedirect(LOGIN_URL)
 
 
 class ReplyCreateView(LoginRequiredMixin, View):
@@ -96,21 +103,25 @@ class ReplyCreateView(LoginRequiredMixin, View):
 @login_required
 def accept_reply(request, pk):
     reply = Reply.objects.get(id=pk)
-    reply.accept = True
-    reply.save()
+    if reply.post.author == request.user:
+        reply.accept = True
+        reply.save()
 
-    send_mail(subject=f'Принят ваш отклик на портале billboard.',
-              message=f'Принят ваш отклик на портале billboard.',
-              from_email='billboard <matoko18@yandex.ru>',
-              recipient_list=[f'{reply.author.email}'],
-              html_message=f'<h2>Ваш отклик на объявление "{reply.post.title}" принят</h2>',
-              fail_silently=True)
+        send_mail(subject=f'Принят ваш отклик на портале billboard.',
+                  message=f'Принят ваш отклик на портале billboard.',
+                  from_email='billboard <matoko18@yandex.ru>',
+                  recipient_list=[f'{reply.author.email}'],
+                  html_message=f'<h2>Ваш отклик на объявление "{reply.post.title}" принят</h2>',
+                  fail_silently=True)
 
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    return HttpResponseRedirect(LOGIN_URL)
 
 
 @login_required
 def delete_reply(request, pk):
     reply = Reply.objects.get(id=pk)
-    reply.delete()
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    if reply.post.author == request.user:
+        reply.delete()
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    return HttpResponseRedirect(LOGIN_URL)
